Added web scrapers for SCDF and BCA circulars
